Remove the emulator for stack size.
