+++ pyelftool_parser/src/analyzer.py
@@ -1,4 +1,5 @@
 import sys
+import disasmbler
 import register
 import execute
 import jmp_class
@@ -13,208 +14,8 @@ from capstone import *
 from elftools.elf.sections import (
     NoteSection, SymbolTableSection, SymbolTableIndexSection
 )
-class disassembler:
-    def __init__(self, path, entry_function):
-        self.path = path
-        self.inst = dict()              ## mapping address to inst encoding
-        self.symbol = dict()            ## mapping address to symbol
-        self.symbol_address = dict()    ## mapping symbol to address
-        self.syn_invocation = dict()    ## mapping the synchronization symbol address to target function
-        self.invo_jmp_table = dict()    ## hardcode the invocation address for fetching pc 
-        self.thread_list = dict()    ## hardcode the call/jmp for fetching pc 
-        self.slm_ipithd_create_address = 0  ## hardcode the thread function for scheduler.
-        self.capmgr_initthd_create_address = 0
-        self.function_call_address = 0
-        
-        self.vertex = dict()            ## construct the calling graph
-        self.entry_function = entry_function
-        self.entry_function_list = list()    ## need to put the cosrt_s into it.
-        self.entry_pc = 0
-        self.exit_pc = 0
-        self.invocation_function = list() ## record the invocation function here.
-        self.acquire_stack_size = 0
-        self.init_done_address = 0
-        self.init_done_end_address = 0
-                
-    def disasmstubs(self, file_path):
-        stub_pattern = re.compile(r'cos_asm_stub\((\w+)\)')
-        stub_indirect_pattern = re.compile(r'cos_asm_stub_indirect\((\w+)\)')
-        stubs = []
-        stub_indirects = []
-
-        # Open and read the file
-        with open(file_path, 'r') as file:
-            content = file.read()
-
-            # Find all cos_asm_stub matches
-            stubs = stub_pattern.findall(content)
-
-            # Find all cos_asm_stub_indirect matches
-            stub_indirects = stub_indirect_pattern.findall(content)
-
-        # Output the results
-        log("cos_asm_stub functions:")
-        for stub in stubs:
-            self.invocation_function.append(stub)
-
-        log("\ncos_asm_stub_indirect functions:")
-        for stub_indirect in stub_indirects:
-            self.invocation_function.append(stub_indirect)
-
-    def disasminstpass(self, md, ops, addr):    ## pinpoint all the needed hardcode address, also decode the insts.
-        pc_flag = 0
-        stack_flag = 0
-        
-        for inst in md.disasm(ops, addr):
-            self.inst[inst.address] = (inst)
-            if (inst.address in self.symbol):   ## it is to catch the next symbol start, also for catch the last inst. @minghwu: I think it could have a better way to do this.
-                pc_flag = 0
-                
-            if (inst.address == self.entry_pc): ## when it is entry point, set the flag = 1 to catch exitpoint.
-                pc_flag = 1
-            if (pc_flag == 1):                  ## catch the point until the next symbol, means it is exit point.
-                self.exit_pc = inst.address
-            
-            if inst.address in self.symbol and self.symbol[inst.address] == "custom_acquire_stack":  ## try to catch the movabs which is in the acquire_stack function to set rsp.
-                stack_flag = 1
-            
-            if stack_flag == 1 and inst.id == X86_INS_MOVABS:
-                stack_flag = 0
-                tempstack = 0
-                for i in inst.operands:
-                    if i.type == X86_OP_IMM:
-                        tempstack = i.imm
-                self.acquire_stack_size = tempstack
-                log("stack_size_begin")
-                log(self.acquire_stack_size)
-    
-    def disasmthreadpointer(self, md, ops, addr):
-        thread_function_list = list()
-        for inst in md.disasm(ops, addr):
-            if inst.address in self.symbol and self.symbol[inst.address] == "slm_ipithd_create": 
-                self.slm_ipithd_create_address = inst.address
-                thread_function_list.append(self.slm_ipithd_create_address)
-            if inst.address in self.symbol and self.symbol[inst.address] == "capmgr_initthd_create": 
-                self.capmgr_initthd_create_address = inst.address
-                thread_function_list.append(self.capmgr_initthd_create_address)                
-        flag = 0
-        for inst in md.disasm(ops, addr):
-            if inst.address in self.symbol and self.symbol[inst.address] == "cos_upcall_fn":
-                flag = 1
-            if flag == 1 and inst.id == X86_INS_CALL:
-                for i in inst.operands:
-                    if i.type != X86_OP_MEM and i.type != X86_OP_IMM: ## mean it is not the call instruction we want, because it is memory call.
-                        if len(thread_function_list) > 0:
-                            self.function_call_address = inst.address
-                            self.thread_list= thread_function_list
-                            flag = 0
-
-    def disasminst(self):  ## decode the inst for execute
-        with open(self.path, 'rb') as f:
-            elf = ELFFile(f)
-            code = elf.get_section_by_name('.text')
-            ops = code.data()
-            addr = code['sh_addr']
-            md = Cs(CS_ARCH_X86, CS_MODE_64)
-            md.detail = True
-            self.disasminstpass(md, ops, addr)  ## disasm the instruction into a list, setting up the entry/exit pc.
-            self.disasmthreadpointer(md, ops, addr) ## hardcode the dynamic thread.
-    def disasmsymbol(self):
-        with open(self.path, 'rb') as f:
-            e = ELFFile(f)
-            symbol_tables = [ s for s in e.iter_sections()
-                         if isinstance(s, SymbolTableSection)]
-            for section in symbol_tables:
-                for symbol in section.iter_symbols():
-                    self.symbol[symbol['st_value']] = symbol.name
-                    self.symbol_address[symbol.name] = symbol['st_value']
-                    self.vertex[symbol['st_value']] = symbol.name
-                    log(symbol.name, symbol['st_value'])
-                    if(symbol.name == self.entry_function): ## set up the entry pc.
-                        self.entry_pc = symbol['st_value']
-                        log("Set up entry point")
-                        log(hex(self.entry_pc))
-                    if(symbol.name == 'custom_acquire_stack'):
-                        self.acquire_stack_address = symbol['st_value']
-                    
-    
-    def disasminvocation(self):  ## change the cosrt_extern to cosrt_c_;
-        with open(self.path, 'rb') as f:
-            e = ELFFile(f)
-            symbol_tables = [ s for s in e.iter_sections()
-                         if isinstance(s, SymbolTableSection)]
-            for section in symbol_tables:
-                for symbol in section.iter_symbols():
-                    if "__cosrt_s" in symbol.name:
-                        if symbol.name.replace("__cosrt_s_", "") in self.invocation_function:
-                            self.entry_function_list.append(symbol.name)
-                    
-                    if "__cosrt_extern" in symbol.name:
-                        if symbol.name.replace("__cosrt_extern_", "") in self.invocation_function:
-                            if symbol.name.replace("__cosrt_extern", "__cosrt_c") in self.symbol.keys():   ## check is their mapping __cosrt_extern_* to __cosrt_c_* invocation
-                                log("invocation1")
-                                self.syn_invocation[symbol['st_value']] =  self.symbol_address[symbol.name.replace("__cosrt_extern", "__cosrt_c")]
-                                log(self.syn_invocation[symbol.name])
-                            else:
-                                log("invocation2")
-                                log(symbol.name)
-                                self.syn_invocation[symbol['st_value']] = self.symbol_address["__cosrt_c_cosrtdefault"]
-                    if "__cosrt_extern_init_done" in symbol.name:
-                        self.init_done_address = symbol['st_value'] 
-                    if "__cosrt_extern_init_parallel_await_init":
-                        self.init_parallel_await_init_address = symbol['st_value']
-
-    def disasminvotable(self):  ## I hardcode synchr. 
-        with open(self.path, 'rb') as f:
-            elf = ELFFile(f)
-            code = elf.get_section_by_name('.text')
-            ops = code.data()
-            addr = code['sh_addr']
-            md = Cs(CS_ARCH_X86, CS_MODE_64)
-            md.detail = True
-            
-            for inst in md.disasm(ops, addr): ## hardcode the call and jmp address here.
-                if inst.address in self.syn_invocation: ## build up the invocation table.
-                    self.invo_jmp_table[inst.address] = self.syn_invocation[inst.address]
-
-
-    def sym_analyzer(self):
-        sym_info = {}
-        with open(self.path, 'rb') as f:
-            e = ELFFile(f)
-            symbol_tables = [ s for s in e.iter_sections()
-                         if isinstance(s, SymbolTableSection)]
-            for section in symbol_tables:
-                for symbol in section.iter_symbols():
-                    if (symbol['st_size'] == 0):
-                        continue   
-                    sym_info[symbol.name] = {
-                        'address': symbol['st_value'],
-                        'size': symbol['st_size'],
-                        'padding': 0   
-                    }
-            #sort by adress to ensure contiguous symbols
-            sorted_names = sorted(sym_info.keys(),
-            key=lambda name: sym_info[name]['address'],
-            reverse=False)
-            #assign padding based on difference in address
-            for i, name in enumerate(sorted_names):
-                if (i == 0):
-                    continue
-                prev_sym = sym_info[sorted_names[i - 1]]
-                cur_sym = sym_info[name]
-                prev_sym['padding'] = cur_sym['address'] - prev_sym['address'] - prev_sym['size']
-                #sort by size
-            sorted_names = sorted(sym_info.keys(),key=lambda name: sym_info[name]['size'], reverse=True)
-            #print symbols in order of size with padding
-            for name in sorted_names[:10]:
-                cur_sym = sym_info[name]
-                log(
-                    f"Name: {name}, Address: {hex(cur_sym['address'])}, Size: {hex(cur_sym['size'])}, Padding: {hex(cur_sym['padding'])}"
-                )
-    
 class parser:
-    def __init__(self, symbol, inst, register, execute, entry_pc, exit_pc, acquire_stack_address, invo_jmp_table, thread_list, function_call_address, basic_block_mode):
+    def __init__(self, symbol, inst, register, execute, entry_pc, exit_pc, acquire_stack_address, invo_jmp_table, thread_list, function_call_address):
         self.symbol = symbol 
         self.inst = inst
         self.stacklist = []
@@ -235,16 +36,18 @@ class parser:
         self.retcallpc = []
         self.seenlist = [] ## handle the while loop jmp.
         self.JtypeClass = []
-        self.basic_block_mode = basic_block_mode
-    def check_virtual_return(self, address_list):
+
+    def check_exe_virtual_return(self, address_list, function_now): ## virtual ret.
         if address_list[self.index + 1] in self.symbol.keys() : ## have a virtual return address for some function does not have ret.
-            self.execute.exe(-1)  ## virtual ret.
+            self.execute.exe(-1)  
             if self.execute.retflag == 1:
                 log("virtual return")
                 return 1
         return 0
+    
     def IsJtypeInst(self, inst):
         return inst.id == X86_INS_CALL or inst.id == X86_INS_JMP or inst.id == X86_INS_JE or inst.id == X86_INS_JLE or inst.id == X86_INS_JGE or inst.id == X86_INS_JG or inst.id == X86_INS_JNE
+    
     def stack_analyzer(self):
         address_list = list(self.inst.keys())  ## a list for instruction address.
         address_list.append(-1) ## dummy value for last iteration.
@@ -252,16 +55,16 @@ class parser:
         nextinstRip = list(self.inst.keys())
         nextinstRip.append(-1) ## dummy value for last iteration.
         self.register.updaterip(nextinstRip[self.index + 1 if self.index + 1 in nextinstRip else self.index]) ## catch the rip for memory instruction.
+        function_now = 0
         while(self.register.reg["pc"] != self.exit_pc):
-            
-            
+            if self.register.reg["pc"] in self.symbol:
+                function_now = self.symbol[self.register.reg["pc"]]
+                
             #### execute 
-            
             self.execute.exe(self.inst[self.register.reg["pc"]])
             self.register.updatesmaxstackreg()
-            #### fetch next instruction pc and commit
-            vertexfrom = address_list[self.index]  ## this is for graph
             
+            #### fetch next instruction pc and commit
             if address_list[self.index] in self.seenlist: ## Already seen this address before, going to context switch to last branch.
                 self.seenlist.remove(address_list[self.index])
                 if len(self.JtypeClass) > 0:
@@ -269,26 +72,31 @@ class parser:
                     self.index = branchnode.returnPCIndex
                     self.register.reg["stack"] = branchnode.stack
                     self.register.reg["rsp"] = branchnode.rsp
-                    self.register.reg["rspbegin"] = branchnode.rspbegin    
-            elif address_list[self.index] in self.invo_jmp_table:  ### hardcode the synchronization table, we jmp to target address.
+                    self.register.reg["rspbegin"] = branchnode.rspbegin
+                    
+            elif address_list[self.index] in self.invo_jmp_table:         ## looking up hardcode the synchronization table, we jmp to target address.
                 if self.inst[address_list[self.index]].id == X86_INS_CALL:
                     self.JtypeClass.append(jmp_class.JmpContext(self.index+1, self.index, self.register.reg["stack"], self.register.reg["rspbegin"], self.register.reg["rsp"]))
                 if self.IsJtypeInst(self.inst[address_list[self.index]]):
                     self.JtypeClass.append(jmp_class.JmpContext(self.index+1, self.index, self.register.reg["stack"], self.register.reg["rspbegin"], self.register.reg["rsp"]))
                 self.seenlist.append(address_list[self.index])
-                self.edge.add((hex(vertexfrom), hex(self.invo_jmp_table[address_list[self.index]])))
+                if self.invo_jmp_table[address_list[self.index]] in self.symbol:
+                    self.edge.add((function_now, self.symbol[self.invo_jmp_table[address_list[self.index]]]))
                 self.index = address_list.index(self.invo_jmp_table[address_list[self.index]])
-                self.register.reg["call_or_jmp"] = 0 ## clean the call/jmp reg. 
+                self.register.reg["call_or_jmp"] = 0 ## clean the call/jmp indicator. 
                 log("fastpace with hardcode invocation table.")
-            elif address_list[self.index] == self.function_call_address:  ### hardcode the thread address.
+
+            elif address_list[self.index] == self.function_call_address:  ## looking up hardcode the thread address, and jmp to target address.
                 self.JtypeClass.append(jmp_class.JmpContext(self.index + 1, self.index, self.register.reg["stack"], self.register.reg["rspbegin"], self.register.reg["rsp"]))
-                for thread_function_address in self.thread_list: ## here is interesting part, self triger self to search
+                for thread_function_address in self.thread_list:
                     self.JtypeClass.append(jmp_class.JmpContext(address_list.index(thread_function_address), self.index, self.register.reg["stack"], self.register.reg["rspbegin"], self.register.reg["rsp"]))
-                    self.edge.add((hex(vertexfrom), thread_function_address))
+                    if thread_function_address in self.symbol:
+                        self.edge.add((function_now, self.symbol[thread_function_address]))
                 self.seenlist.append(address_list[self.index])
-                self.register.reg["call_or_jmp"] = 0 ## clean the call/jmp reg. 
+                self.register.reg["call_or_jmp"] = 0     ## clean the call/jmp indicator. 
                 log("fastpace with hardcode thread table.")
-            elif self.register.reg["call_or_jmp"] == 0:  ## it is not jmp/call inst, try to fetch next instruction
+
+            elif self.register.reg["call_or_jmp"] == 0:  ## if it is not jmp/call inst, try to fetch next instruction
                 if self.inst[self.register.reg["pc"]].id == (X86_INS_RET): ## ret instruction, go to return address.
                     if len(self.JtypeClass) > 0:
                             branchnode = self.JtypeClass.pop()
@@ -297,7 +105,7 @@ class parser:
                             self.register.reg["rsp"] = branchnode.rsp
                             self.register.reg["rspbegin"] = branchnode.rspbegin
                 else:
-                    if self.check_virtual_return(address_list):
+                    if self.check_exe_virtual_return(address_list, function_now):
                         if len(self.JtypeClass) > 0:
                             branchnode = self.JtypeClass.pop()
                             self.index = branchnode.returnPCIndex
@@ -311,7 +119,7 @@ class parser:
                     if self.register.reg["call_or_jmp"] == 2:  ## handle unknown function pointer.
                         logerror("Here is dynamic call")
                         logerror(address_list[self.index], self.inst[address_list[self.index]].mnemonic, self.inst[address_list[self.index]].op_str)
-                        if self.check_virtual_return(address_list):
+                        if self.check_exe_virtual_return(address_list, function_now):
                             if len(self.JtypeClass) > 0:
                                 branchnode = self.JtypeClass.pop()
                                 self.index = branchnode.returnPCIndex
@@ -319,13 +127,14 @@ class parser:
                                 self.register.reg["rsp"] = branchnode.rsp
                                 self.register.reg["rspbegin"] = branchnode.rspbegin
                         else:
-                            self.index = self.index + 1 
+                            self.index = self.index + 1                             
                     elif address_list[self.index] not in self.seenlist:
-                        # self.retcallpc.append(self.inst[address_list[self.index + 1]].address)
                         self.JtypeClass.append(jmp_class.JmpContext(self.index+1, self.index, self.register.reg["stack"], self.register.reg["rspbegin"], self.register.reg["rsp"]))
                         self.seenlist.append(address_list[self.index])
-                        self.edge.add((hex(vertexfrom), hex(self.register.reg["pc"])))
+                        if self.register.reg["pc"] in self.symbol:
+                            self.edge.add((function_now, self.symbol[self.register.reg["pc"]]))
                         self.index = address_list.index(self.register.reg["pc"])
+                        
                     else: ## It is seen, time to pop.
                         self.seenlist.remove(address_list[self.index])
                         if len(self.JtypeClass) > 0:
@@ -339,7 +148,7 @@ class parser:
                     if self.register.reg["call_or_jmp"] == 2:  ## unknown function pointer or already seen
                         logerror("Here is dynamic jmp")
                         logerror(address_list[self.index], self.inst[address_list[self.index]].mnemonic, self.inst[address_list[self.index]].op_str)
-                        if self.check_virtual_return(address_list):
+                        if self.check_exe_virtual_return(address_list, function_now):
                             if len(self.JtypeClass) > 0:
                                 branchnode = self.JtypeClass.pop()
                                 self.index = branchnode.returnPCIndex
@@ -351,7 +160,8 @@ class parser:
                     elif address_list[self.index] not in self.seenlist:
                         self.seenlist.append(address_list[self.index])
                         self.JtypeClass.append(jmp_class.JmpContext(self.index+1, self.index, self.register.reg["stack"], self.register.reg["rspbegin"], self.register.reg["rsp"]))
-                        self.edge.add((hex(vertexfrom), hex(self.register.reg["pc"])))
+                        if self.register.reg["pc"] in self.symbol:
+                            self.edge.add((function_now, self.symbol[self.register.reg["pc"]]))
                         self.index = address_list.index(self.register.reg["pc"])
                     else: ## it is seen, time to pop.
                         self.seenlist.remove(address_list[self.index])
@@ -360,11 +170,9 @@ class parser:
                             self.index = branchnode.returnPCIndex
                             self.register.reg["stack"] = branchnode.stack
                             self.register.reg["rsp"] = branchnode.rsp
-                            self.register.reg["rspbegin"] = branchnode.rspbegin
-                            
+                            self.register.reg["rspbegin"] = branchnode.rspbegin               
             
             #### commit the instruction.
-            
             self.register.reg["call_or_jmp"] = 0 ## clean the call/jmp reg.
             self.register.updatesmaxstackreg()
             self.register.reg["pc"] = address_list[self.index]
@@ -375,33 +183,31 @@ class parser:
             ########
         return (self.stackfunction,self.stacklist)
 class driver:
-    def __init__(self, path, entry_function, stub_path, basic_block_mode) -> None:
-        self.disassembler = disassembler(path, entry_function)
+    def __init__(self, path, entry_function, stub_path) -> None:
+        self.disasmbler = disasmbler.disasmbler(path, entry_function)
         if os.path.exists(stub_path):
-            self.disassembler.disasmstubs(stub_path)
-        self.disassembler.disasmsymbol()
-        self.disassembler.disasminvocation()  ##TODO @minghwu we also need to consider the cosrt_s_ from here as entry point.
-        self.disassembler.disasminst()
-        self.disassembler.disasminvotable()
-
-        self.disassembler.sym_analyzer()
-        log("program entry:"+ str(self.disassembler.entry_pc))
-        log("program exit:"+ str(self.disassembler.exit_pc))
-        log("program stacksize"+ str(self.disassembler.acquire_stack_size))
-        self.register = register.register(self.disassembler.acquire_stack_size)
-        self.register.reg["pc"] = self.disassembler.entry_pc
+            self.disasmbler.disasmstubs(stub_path)
+        self.disasmbler.disasmsymbol()
+        self.disasmbler.disasminvocation()  ##TODO @minghwu we also need to consider the cosrt_s_ from here as entry point.
+        self.disasmbler.disasminst()
+        self.disasmbler.disasminvotable()
+        
+        log("program entry:"+ str(self.disasmbler.entry_pc))
+        log("program exit:"+ str(self.disasmbler.exit_pc))
+        log("program stacksize"+ str(self.disasmbler.acquire_stack_size))
+        self.register = register.register(self.disasmbler.acquire_stack_size)
+        self.register.reg["pc"] = self.disasmbler.entry_pc
         self.execute = execute.execute(self.register, )
-        self.parser = parser(self.disassembler.symbol, 
-                        self.disassembler.inst, 
+        self.parser = parser(self.disasmbler.symbol, 
+                        self.disasmbler.inst, 
                         self.register,
                         self.execute,
-                        self.disassembler.entry_pc,
-                        self.disassembler.exit_pc, 
-                        self.disassembler.acquire_stack_address,
-                        self.disassembler.invo_jmp_table,
-                        self.disassembler.thread_list,
-                        self.disassembler.function_call_address,
-                        basic_block_mode)
+                        self.disasmbler.entry_pc,
+                        self.disasmbler.exit_pc, 
+                        self.disasmbler.acquire_stack_address,
+                        self.disasmbler.invo_jmp_table,
+                        self.disasmbler.thread_list,
+                        self.disasmbler.function_call_address)
     
     def PowerOf2(self, N):
         # Calculate log2 of N
@@ -417,20 +223,10 @@ class driver:
         self.parser.stack_analyzer()
         
         logresult(self.parser.edge)
-        
-
-        for j in zip(self.parser.stackfunction, self.parser.stacklist):
-            logresult(j)
-            
         logresult(self.register.reg["max"])
         logrust(self.PowerOf2(abs(self.register.reg["max"])))        
 
 if __name__ == '__main__':
-    
-    ## path = "../testbench/composite/system_binaries/cos_build-test/global.sched/sched.pfprr_quantum_static.global.sched"
-    ## path = "/home/minghwu/work/minghwu/composite/system_binaries/cos_build-test/global.ping/tests.unit_pingpong.global.ping"   
-    
-    basic_block_mode = 0  ## Is it basic block mode = 1 / stack mode = 0.
     
     if len(sys.argv) >=3:
         entry_function = sys.argv[2]
@@ -444,6 +240,6 @@ if __name__ == '__main__':
         stub_path = "../../src/components/interface/" + "pong" + "/stubs/stubs.S"
     else:
         stub_path = "../../src/components/interface/" + path.split(".")[-1] + "/stubs/stubs.S"
-    driver = driver(path, entry_function, stub_path, basic_block_mode)
+    driver = driver(path, entry_function, stub_path)
     driver.run()
